refactor(codemode): split provider implementations
